feat: add get hot scholars

+++ papers/models.py
@@ -217,10 +217,10 @@ class AreaStatistics(models.Model):
 
 class PaperRank(models.Model):
     pid = models.BigIntegerField(primary_key=True)
-    rank = models.IntegerField(default=0)
+    rank = models.FloatField(default=0.0)
 
     @classmethod
-    def create(cls, _pid, _rank=0):
+    def create(cls, _pid, _rank=0.0):
         return cls(pid=_pid, rank=_rank)
 
     class Meta:
